Refactor Merchandise and shelves
Shelves are now saved in a special section in the InteriorRefList and must be loaded dynamically when the shop is loaded.

Merchandise for the shop is loaded once and saved in the private merchandise chest for the shop. Optimize updating merchandise to refresh the shelves from the response of the update instead of refetching the merchandise.

A single shelf can be loaded and paged indpedent from other shelves in the cell.